% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARI.R
\name{adjRand}
\alias{adjRand}
\title{Adjusted Rand Index}
\usage{
adjRand(class1, class2)
}
\arguments{
\item{class1}{A \eqn{n}-vector of class labels.}

\item{class2}{A \eqn{n}-vector of class labels.}
}
\value{
A scalar value, taking the value 1 if the two classifications are
in perfect agreement. Under random classification, the adjusted Rand Index is 0.
}
\description{
Adjusted Rand Index
}
\details{
Compute the adjusted Rand Index comparing the two 
classifications, \code{class1} and \code{class2}.
}
\references{
Lawrence Hubert & Phipps Arabie, 1985. "Comparing partitions," Journal of
Classification, 2(1): 193-218.
}
\author{
Emily Goren, \email{emily.goren@gmail.com}
}
