% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\title{Classify New Observations}
\usage{
classify(newobs, params)
}
\arguments{
\item{newobs}{A matrix with new observations (rows), \eqn{p} columns
(dimensions), and missing entries set to \code{NA}.}

\item{params}{A list of parmaters defining a finite mixture of
multivariate t distributions, 
usually obtained from \code{\link{MixtClust}} (see details).}
}
\value{
A vector classify each observation to the cluster \eqn{1, \dots, K}
with the highest posterior probability.
}
\description{
Classify new, possibly incomplete, observations arising from a finite mixture of multivariate
t distributions.
}
\details{
Classify new observations according to the finite mixture of t distirbutions
specified by the parameter values in \eqn{params}, a named list with elements:
\itemize{ 
  \item{"pi"}{Mixing proportions. A vector of length \eqn{K} that
sums to one.} 
  \item{"nu"}{Degrees of freedom. A vector of length \eqn{K} with
entries at least equal to three (thus requiring the existance of the first
two moments.)} 
  \item{"mu"}{Locations. A \eqn{K \times p} matrix, where the
\eqn{k}-th row is the location \eqn{\mu_k \in R^p} for cluster
\eqn{k}.} 
  \item{"Sigma"}{Dispersions. A \eqn{p \times p \times K} array, where
the \eqn{k}-th slice is the \eqn{p \times p} positive-definite disperion
matrix \eqn{\Sigma_k} for cluster \eqn{k}.} }
}
\author{
Emily Goren, \email{emily.goren@gmail.com} based on modifications of
code by Ranjan Maitra.
}
