% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_bic}
\alias{extract_bic}
\title{Extract BIC for list of MixtClust results}
\usage{
extract_bic(res)
}
\arguments{
\item{res}{A list of results given as the output to \code{MixtClust()}}
}
\value{
A data frame with the columns "constraint", "nclusters", and "BIC"
}
\description{
After obtaining a list of results for various constraints and
number of clusters, extract a data frame contaning the BIC for each model.
}
\examples{
set.seed(20241029)
d <- subset(iris, select = -Species)
res <- MixtClust(d, nclusters = 2:3, sigma.constr = "all")
BICs <- extract_bic(res)

}
\author{
Josh Berlinski
}
