% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grb}
\alias{grb}
\title{Gamma ray burst (GRB) data}
\format{
An object of class \code{"matrix"} with 
1973 rows (records/observations) and 9 columns (features).
}
\usage{
data(grb)
}
\description{
Data on gamma ray bursts from the 
Burst and Transient Source Experiment (BATSE) 4Br catalogue. 
There are nine features 
(T50, T90, F1, F2, F3, F4, P64, P256, P1024) on 1,973 GRBs,
where the features are defined as:
\itemize{
  \item{T50, T90:}{The time by which 50 and 90 percent of the flux arrive, respectively.}
  \item{F1, F2, F3, F4:}{The time-integrated fluences in spectral channels 20–50, 50–100, 100–300 and >300 keV, respectively.}
  \item{P64, P256, P1024:}{The peak fluxes as measured within bins of 64, 256 and 1024 ms, respectively.}
}
}
\details{
Missing values are represented with NA and non-missing values 
have been log10 transformed.
}
\examples{
data(grb)

}
\references{
Souradeep Chattopadhyay & Ranjan Maitra, 2017.
"Gaussian-mixture-model-based cluster analysis finds five kinds 
of gamma-ray bursts in the BATSE catalogue," Monthly Notices 
of the Royal Astronomical Society, 
469(3): 3374-3389. 
https://doi.org/10.1093/mnras/stx1024

Souradeep Chattopadhyay & Ranjan Maitra, 2018.
"Multivariate t-mixture-model-based cluster analysis of BATSE 
catalogue establishes importance of all observed parameters,
confirms five distinct ellipsoidal sub-populations of gamma-ray
bursts," Monthly Notices of the Royal Astronomical Society,
481(3): 3196-3209.
https://doi.org/10.1093/mnras/sty1940

Emily M. Goren & Ranjan Maitra, 2022. "Fast model-based clustering of partial records," Stat, 11( 1), e416. https://doi.org/10.1002/sta4.416 
}
\keyword{datasets}
