% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mahalanobis}
\alias{mahalanobis}
\title{Mahalanobis distance}
\usage{
mahalanobis(x, mu, sigma, ischol = FALSE)
}
\arguments{
\item{x}{Numeric. A vector (of length \eqn{p}) or matrix (with \eqn{p}
columns).}

\item{mu}{Numeric. A vector of length \eqn{p}.}

\item{sigma}{Numeric. A \eqn{p \times p} non-negative definite matrix.}

\item{ischol}{Logical. Set to \code{TRUE} if \code{sigma} is provided as a
Cholesky decomposition.}
}
\value{
The squared Mahalanobis distance for all rows in \code{x} and the
  mean vector \code{mu} with respect to covariance matrix \code{sigma},
  defined as \eqn{(x - \mu)' \Sigma^{-1}(x - \mu)}.
}
\description{
Compute the squared Mahalanobis distance (fast implementation).
}
\references{
Mahalanobis, Prasanta Chandra. "On the Generalised Distance in
  Statistics." Proceedings of the National Institute of Sciences of India 2
  (1936): 49-55.
}
\author{
Emily Goren, \email{emily.goren@gmail.com}
}
