% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixtClust-package.R
\docType{package}
\name{MixtClust-package}
\alias{MixtClust-package}
\alias{_PACKAGE}
\title{MixtClust: Robust Clustering for Complete and Incomplete Data}
\description{
Model-based clustering using finite mixtures of t distributions, with handling of incomplete data using either marginalization or the EM algorithm.
}
\author{
\strong{Maintainer}: Emily Goren \email{emily.goren@gmail.com}

Authors:
\itemize{
  \item Ranjan Maitra \email{maitra@iastate.edu}
  \item Joshua Berlinski \email{jberlin@iastate.edu}
}

}
\keyword{internal}
